Include zint sources in gem file and build it at 'gem install'
If libzint is already installed on the system and if it's Version 2.10.x, then this version is preferred.
Otherwise the bundled version is built and used.
System or bundled version can be forced by using:
  gem install ruby-zint -- --enable-system-libzint  or  --disable-system-libzint

This adds mini_portile2 as a helper to download and build zint.
The build requires installed cmake and libpng.
A receipt for libpng might be added to this project as well, but for now it is expected on the system.