Move FFI functions to Zint::Native and improve structs
This moves the FFI functions to the private module Zint::Native, to hide it as internal API.

Make use of Struct.by_ref to pass and retrieve Struct classes instead of plain pointers.
They are type checked that way and it's less ruby code.

Fix memory leak of Struct::Symbol and attached structures not being GC'ed.

Add proper methods for retrieval of barcode vector parts instead of a plain FFI::Struct.