Allow only one export per barcode
libzint allows to export several barcodes through the same Symbol structure and frees memory associated to the previous barcode before exporting the new one.

In Ruby this is problematic, since it makes safe memory handling very difficult.
Previous exported vectors would crash with segfaults, if they are accessed.

To overcome this situation a tracking of previously exported vectors would have to be done, which complicates the sources.

The alternative is to not allow multiple exports with one and the same Symbol structure.
This is what this commit implements.

The Barcode object is still a mutable object, but it is preferred to pass all parameters to Barcode.new .
Multiple exports are answered with Zint::Barcode::AlreadyGenerated .

Also change spec to use the preferred method of passing all values to Barcode.new .