Add cross compiled native gems for various platforms
rake-compiler is made to build ruby extensions, but we need a DLL which is loaded by FFI.
The DLL is not dependent on the ruby version, other than a ruby extension.
Therefore this patch makes some tricks to reuse rake-compiler.

1. Our libzint file is named according to the needs of rake-compiler
2. RUBY_CC_VERSION is set to 3.4.1:0.0.0 to get one compiler run and to work around dependency issues with native lib/libzint.so.
3. The runtime path for libzint is ajdusted to rake-compilers path standards

Native binary gems can be built by:

```
rake gem:x86_64-linux-gnu  # to build the gem for a specific platform
rake gem:native            # to build all gems for all platforms in parallel
```